requirements_compress.txt
Pillow==10.2.0
numpy==1.26.3